update ovphysx 0.5.10

+++ physx/source/physxextensions/src/serialization/Xml/PsFastXml.h
@@ -22,7 +22,7 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
-// Copyright (c) 2008-2025 NVIDIA Corporation. All rights reserved.
+// Copyright (c) 2008-2026 NVIDIA Corporation. All rights reserved.
 // Copyright (c) 2004-2008 AGEIA Technologies, Inc. All rights reserved.
 // Copyright (c) 2001-2004 NovodeX AG. All rights reserved.
 
@@ -125,12 +125,12 @@ class FastXml
 		virtual bool processDoctype(const char* /*rootElement*/, // Root element tag
 		                            const char* /*type*/,        // SYSTEM or PUBLIC
 		                            const char* /*fpi*/,         // Formal Public Identifier
-		                            const char* /*uri*/)         // Path to schema file
+		                            const char* /*uri*/) // Path to schema file
 		{
 			return true;
 		}
 
-		virtual void* allocate(uint32_t size)
+		virtual void* allocate(uint64_t size)
 		{
 			return PxGetBroadcastAllocator()->allocate(size, "FastXml", PX_FL);
 		}
